requirements.txt
torch
torchvision
pillow
transformers
clip
scipy
numpy
matplotlib
tqdm
